move Spot to weak depends
