+++ src/optim/discrete_systems/cosafe_ltl_problem.jl
@@ -2,8 +2,6 @@
 # CoSafe LTL Control
 # ============================================================
 
-import Spot
-
 mutable struct OptimizerCoSafeLTLProblem{T} <: MOI.AbstractOptimizer
     # inputs
     problem::Union{Nothing, PR.CoSafeLTLProblem}
@@ -63,7 +61,7 @@ function MOI.optimize!(optimizer::OptimizerCoSafeLTLProblem)
     problem === nothing && error("problem not set")
     _check_ap_coverage(problem)
 
-    if optimizer.print_level > 0
+    if optimizer.print_level > 0 && problem.labeling isa AbstractDict
         for (ap, states) in problem.labeling
             println("AP=$ap  #states=$(length(states))")
         end
@@ -72,13 +70,11 @@ function MOI.optimize!(optimizer::OptimizerCoSafeLTLProblem)
     # (1) Finite system
     autom = problem.system
 
-    # (2) Specification automaton
-    spec = if problem.spec isa Spot.SpotFormula
-        spot_stepper(problem.spec)
-    elseif problem.spec isa AbstractSpecStepper
+    # (2) Specification automaton / monitor
+    spec = if problem.spec isa AbstractSpecStepper
         problem.spec
     else
-        error("spec must be SpotFormula or AbstractSpecStepper")
+        error("problem.spec must be an AbstractSpecStepper")
     end
 
     labeling = if problem.labeling isa Function
@@ -109,10 +105,6 @@ function MOI.optimize!(optimizer::OptimizerCoSafeLTLProblem)
     isempty(product_initial_set) && error("Empty product initial set.")
 
     accQ = accepting_states(spec)
-    labels_seen = Set{Any}()
-    for qs in 1:ST.get_n_state(autom)
-        push!(labels_seen, labeling(qs))
-    end
     product_target_set =
         [p for p in 1:ST.get_n_state(product_autom) if product_autom.rev[p][2] in accQ]
     isempty(product_target_set) &&
@@ -204,29 +196,24 @@ end
 # Helpers
 # ============================================================
 
-# Small check: if the spec is a Spot formula, ensure all APs mentioned in the formula
-# have a set provided in `concrete_problem.labeling`.
-# Extra keys in the dict are fine.)
-collect_aps(φ::Spot.SpotFormula) = [Symbol(ap) for ap in Spot.atomic_prop_collect(φ)]
-
 function _check_ap_coverage(problem)
-    if !(problem.spec isa Spot.SpotFormula)
+    if !(problem.labeling isa AbstractDict)
         return nothing
     end
 
-    problem.labeling isa AbstractDict || return nothing
-
-    aps = Set(collect_aps(problem.spec))
-    provided = Set(keys(problem.labeling))
-    missing = setdiff(aps, provided)
-    if !isempty(missing)
-        error("Missing AP sets in problem.labeling for: $(collect(missing)).")
+    if problem.spec isa FunctionMonitor && !isempty(problem.spec.aps)
+        provided = Set(keys(problem.labeling))
+        missing = setdiff(Set(problem.spec.aps), provided)
+        if !isempty(missing)
+            error("Missing AP sets in problem.labeling for: $(collect(missing)).")
+        end
     end
+
     return nothing
 end
 
 # ============================================================
-# Spec interface (either Spot DRA or user-defined monitor)
+# Spec interface (user-defined monitor)
 # ============================================================
 
 abstract type AbstractSpecStepper end
@@ -239,7 +226,6 @@ step(::AbstractSpecStepper, ::Int, ::Tuple{Vararg{Symbol}}) = error("step not im
 init_state(::AbstractSpecStepper) = 1
 
 # Optional: return the set of "done/accepting" spec states for co-safe reduction.
-# If not provided, we will attempt to compute it (Spot) or require the user to provide.
 accepting_states(::AbstractSpecStepper) = error("accepting_states not implemented")
 
 # --------------------------
@@ -248,116 +234,18 @@ accepting_states(::AbstractSpecStepper) = error("accepting_states not implemente
 struct FunctionMonitor <: AbstractSpecStepper
     qa0::Int
     acc::Set{Int}
-    stepfun::Function  # (qa::Int, ap::Tuple) -> qa2::Int
+    stepfun::Function              # (qa::Int, ap::Tuple) -> qa2::Int
+    aps::Vector{Symbol}
 end
+
+FunctionMonitor(qa0::Int, acc::Set{Int}, stepfun::Function) =
+    FunctionMonitor(qa0, acc, stepfun, Symbol[])
 
 step(M::FunctionMonitor, qa::Int, ap::Tuple{Vararg{Symbol}}) = M.stepfun(qa, ap)
 init_state(M::FunctionMonitor) = M.qa0
 accepting_states(M::FunctionMonitor) = M.acc
 
-# --------------------------
-# Spot-based stepper (DRA used as a monitor)
-# --------------------------
-struct SpotDRAstepper <: AbstractSpecStepper
-    φ::Spot.SpotFormula
-    dra::Any
-    qa0::Int
-    qa_dead::Int
-    aps::Vector{Symbol}  # AP universe used for done-state detection
-end
-
-init_state(S::SpotDRAstepper) = S.qa0
-
-@inline function _nextstate_int(dra, qa::Int, ap::Tuple, qa_dead::Int)
-    qa2 = Spot.nextstate(dra, qa, ap)
-    return qa2 === nothing ? qa_dead : qa2
-end
-
-step(S::SpotDRAstepper, qa::Int, ap::Tuple{Vararg{Symbol}}) =
-    (qa == S.qa_dead) ? S.qa_dead : _nextstate_int(S.dra, qa, ap, S.qa_dead)
-
-# Heuristic done-state detector for Spot DRA used as a co-safe monitor.
-# A state q is considered "done" if for all valuations v:
-#   nextstate(dra, q, v) is either q OR nothing.
-# This is a practical criterion that works well for many co-safe task specs.
-function _all_valuations(aps::Vector{Symbol})
-    n = length(aps)
-    vals = Vector{Tuple{Vararg{Symbol}}}(undef, 1 << n)
-    for mask in 0:((1 << n) - 1)
-        trues = Symbol[]
-        for (i, a) in enumerate(aps)
-            if (mask >> (i-1)) & 1 == 1
-                push!(trues, a)
-            end
-        end
-        vals[mask + 1] = Tuple(trues)
-    end
-    return vals
-end
-
-function _cosafe_done_states_dra(dra; aps::Vector{Symbol}, q0::Int = 1)
-    vals = _all_valuations(aps)
-
-    # reachable spec states (ignore `nothing`)
-    reachable = Set{Int}([q0])
-    queue = [q0]
-    while !isempty(queue)
-        q = popfirst!(queue)
-        for v in vals
-            q2 = Spot.nextstate(dra, q, v)
-            q2 === nothing && continue
-            if !(q2 in reachable)
-                push!(reachable, q2)
-                push!(queue, q2)
-            end
-        end
-    end
-
-    done = Set{Int}()
-    for q in reachable
-        ok = true
-        for v in vals
-            q2 = Spot.nextstate(dra, q, v)
-            if !(q2 === nothing || q2 == q)
-                ok = false
-                break
-            end
-        end
-        ok && push!(done, q)
-    end
-    return done, reachable
-end
-
-accepting_states(S::SpotDRAstepper) = begin
-    doneQ, _ = _cosafe_done_states_dra(S.dra; aps = S.aps, q0 = S.qa0)
-    isempty(doneQ) && error(
-        "Could not find any 'done' spec states with the co-safe heuristic. " *
-        "The formula may not be co-safe for this pipeline, or the AP set is mismatched. " *
-        "Try providing a FunctionMonitor with explicit accepting states.",
-    )
-    doneQ
-end
-
-function spot_stepper(
-    φ::Spot.SpotFormula;
-    qa0::Int = 1,
-    qa_dead::Int = 0,
-    aps::Union{Nothing, Vector{Symbol}} = nothing,
-)
-    aps_use = aps === nothing ? collect_aps(φ) : aps
-    try
-        dra = Spot.DeterministicRabinAutomata(φ)
-        return SpotDRAstepper(φ, dra, qa0, qa_dead, aps_use)
-    catch
-        error(
-            "Spot could not provide a DeterministicRabinAutomata with nextstate(). " *
-            "Provide a FunctionMonitor wrapper.",
-        )
-    end
-end
-
 function labeling_function_from_state_sets(lab_abs::Dict{Symbol, <:AbstractVector{Int}})
-    # speed: BitSet membership is cheap
     lab_bits = Dict{Symbol, BitSet}()
     for (ap, states) in lab_abs
         lab_bits[ap] = BitSet(states)
@@ -423,7 +311,6 @@ function build_product_automaton(
             return length(rev)
         end
 
-    # BFS
     work = Int[]
     inqueue = BitSet()
     for qs in initial_set
@@ -436,7 +323,7 @@ function build_product_automaton(
 
     i = 1
     while i <= length(work)
-        p = work[i];
+        p = work[i]
         i += 1
         (qs, qa) = rev[p]
         for u in 1:nU
@@ -445,7 +332,7 @@ function build_product_automaton(
                 qa2 = step(spec, qa, ap)
                 p2 = getpid(qs2, qa2)
                 if !(p2 in inqueue)
-                    push!(work, p2);
+                    push!(work, p2)
                     push!(inqueue, p2)
                 end
             end
@@ -470,6 +357,7 @@ function build_product_automaton(
             post_tab[p][u] = unique(succs)
         end
     end
+
     return ProductAutomaton(sys, labeling, spec, pid, rev, post_tab, pre_tab, nU)
 end
 
@@ -479,7 +367,6 @@ function build_fm_controller_ms(
     contrP::MS.AbstractMap,
     pid::Dict{Tuple{Int, Int}, Int},
 )
-    # fast labeling_qs(qs) -> Tuple{Symbol,...}
     lab_bits = Dict{Symbol, BitSet}((ap => BitSet(states)) for (ap, states) in lab_abs)
     aps = collect(keys(lab_bits))
 
@@ -492,22 +379,18 @@ function build_fm_controller_ms(
     end
 
     qa0 = init_state(spec)
-
     kP = contrP.h
 
-    # output map: (qa, qs) -> u_sym (or Vector) or nothing
     h = function (qa::Int, qs::Int)
         p = get(pid, (qs, qa), nothing)
         p === nothing && return nothing
         return kP(p)
     end
 
-    # memory update: (qa, qs_for_update) -> qa_next
     g = function (qa::Int, qs_for_update::Int)
         return step(spec, qa, labeling_qs(qs_for_update))
     end
 
-    # Domain for output map: defined iff pid exists and product-controller defined
     isdef_qaqs = function (qaqs)
         qa, qs = qaqs
         p = get(pid, (qs, qa), nothing)
@@ -519,8 +402,6 @@ function build_fm_controller_ms(
     end
     X_qaqs = PredicateDomain(isdef_qaqs)
 
-    # Build MS objects
-    # outputmap expects input (qa,qs); we pack into a 2-tuple input for ConstrainedBlackBoxMap
     outmap = MS.ConstrainedBlackBoxMap(2, 1, qaqs -> begin
         qa, qs = qaqs
         h(qa, qs)
