+++ src/problem/problems.jl
@@ -88,7 +88,7 @@ Encodes a **co-safe LTL control problem**.
 
 - `S`: The system to control.
 - `XI`: The initial set of states.
-- `SPEC`: The co-safe LTL specification object (e.g. a Spot/LTL formula or an automaton/monitor wrapper).
+- `SPEC`: The co-safe LTL specification object (an automaton/monitor wrapper).
 - `LAB`: The labeling payload type used in `labeling` (typically a concrete set type such as a LazySet,
          or an abstract labeling such as `Vector{Int}` / bitset / etc.).
 
@@ -111,9 +111,6 @@ Encodes a **co-safe LTL control problem**.
   Per-AP semantics used when converting set labels to abstract labels.
   Values: `Dionysos.Domain.INNER` or `Dionysos.Domain.OUTER`.
 
-- `strict_spot::Bool`:
-  If `true`, enforce Spot-style strict AP semantics / alphabet handling (useful to catch missing APs).
-  If `false`, be permissive (e.g. treat missing APs as false / ignore absent labels depending on your pipeline).
 
 This problem aims to synthesize a controller such that the generated trajectory satisfies the co-safe LTL
 formula, i.e. it reaches an accepting condition in finite time.
@@ -127,7 +124,6 @@ mutable struct CoSafeLTLProblem{S, XI, SPEC, LAB} <: ProblemType
     labeling::Dict{Symbol, LAB}   # Symbol => LazySet (concrete) or Vector{Int} (abstract)
 
     ap_semantics::Dict{Symbol, Any}  # Symbol => DO.INNER / DO.OUTER
-    strict_spot::Bool
 end
 
 struct Infinity <: Real end
